fix: remove outdated cover block styles
We had added these styles when we had a cover block on the home page and laid some text over it. That's no longer there. The styles don't look good on other cover blocks. And there are new, much more flexible ways of creating a similar effect with group blocks that were introduced in Gutenberg in the meantime.