feat: leverage theme.json
I added in some of the available features we could use. This version of theme.json didn't exist when we created this site and these options weren't available. This is an early commit as I think there's more we can do here. Will open a issue for it.

reference: https://developer.wordpress.org/themes/global-settings-and-styles/settings/