styling links
fix: remove underlines in clip path style headers

Occasionally people add links to the clip path headers and the underline looks bad. So I removed it and added underline on hover.

feat: styled links in .entry .entry-content

Links are still underlined, but dotted so it's a little less in your face. On hover they are underlined solid.

I think this addresses #230