fix: pullquote colors
This makes sure the font color is correct in the editor and the front end whether with default styling or altered in the block editor.

fix: pullquote simpler without is-style-solid-color

removing that old class from the database

This old class was throwing a monkey wrench in things. Got rid of it and will wipe it from the database.